Font: added open/read alt classmethod, __init__ now kw_only
we break with the robofab/defcon tradition that Font has a single
positional parameter which is the filesystem path.

Our Font class is pure data, can be constructed from its elementary
components and is fully constructed after __init__ is run.

It can also be initialized from either a path or from a UFOReader
object, using respectively the open and read alternative constructors
(@classmethod).

When lazy=True, the Font maintains a reference to the UFOReader, and
does not load all the glyphs in all the layers, until requested. Same
for the data and images. The rest of the simple plist data is laoded
immediately regardless of 'lazy' option, because they are relatively
small and the cost of adding laziness to them exceeds (in my opinion)
the gains. I may change my mind again after some serious profiling.

Font can be used as a context manager, so they close their reader
upon exiting (if they were loaded lazily, otherwise they have no
reader and close method is no-op).

The Font not only has a 'read' constructor that takes a UFOReader,
but also has a 'write' method that takes a UFOWriter directly, and
can be used in alternative to the regular 'save' method (which as
usual takes a path).

The 'save' method has a overwrite=False option to force overwriting
when doing a saveAs operation (i.e. when explicitly passing a path).

Lots of changes, still playing with the API so I'll forsee even more
in the future.