store 'default-ness' in Layer.default, set LayerSet.defaultLayer from that
This makes it easier to initialize a new LayerSet from an iterable of
Layers. Also we no longer need to override Font's unstructure/structure
methods, nor add a spurious "defaultLayerName" key to the font dict.
The LayerSet constructor and defaultLayer property ensures the invariant
that only one layer in the set is marked as default.
The 'public.default' is always default by definition.