glyph: make image attribute optional
AFAIK, this is the most unused of all glyph attributes, it's also marked optional in the UFO spec. fontTools.ufoLib happily skips it when it is None. Having all Glyph objects carry an unused image object seems like a waste. If one wants to use images, one can set one.